Move tarballs due to the previous Friday
Each event date is calculated by the week it occurs on, so it is assumed
to occur on a Monday. Since the release happens on day 3 of the week, in
order to move tarballs due to the previous Friday, we have to subtract 3
days from the date. The release event still occurs on date + 2 because
it still happens on Wednesday.