Delay release reminders one extra day
I think we're sending out release reminders a bit too soon. If we send
these Tuesday night, developers are liable to put it off until later in
the week and forget. Let's send the reminders Wednesday night instead,
which gives developers Thursday, Friday, and Saturday to do the release
after the reminder is sent. That's still a reasonable amount of time.