Sleep one second when receiving HTTP 429 error
One of our mirrors has started rate-limiting us. It seems this mirror is
selected for me 100% of the time, and tarball conversion fails 100% of
the time. Ouch. We can avoid this by sleeping for one second and
retrying when it occurs.

Note this is called on a thread, so it theoretically only affects the
thread that actaully received the 429 response. This means that other
threads are likely to hit the same error, but that's OK.

Fixes #23