Fix the freeze announcement code
In GNOME 3.4 we added a "the-freeze" category but failed to give it a
description. We also stopped using previous categories that still lived
on in this script. Remove them.

Also, always print the date that a freeze begins, to avoid confusion,
since the announcement mails are sent prior to the start of the freeze.