42: hold gtk+-3 at 3.24.35
GTK+ 3.24.36 requires meson 0.60 but only 0.59.4 is available
in our 42 build environment.

It requires more work than just downgrading the meson check to 0.59,
so sticking with 3.24.35 seems the simplest solution.

3.24.34 may be a better choice than 3.24.35 because of some regressions;
however, we are unaware of GTK3 complaints from our GNOME 42.7 release.
And it seems more sensible to not backtrack versions if it's not needed.