Add GNOME 44 limits to tarball-conversion and fix obvious problems
Remove modules in tarball-conversion that don't exist in core-deps

Add limits for GNOME 44. I was aggressive adding limits wherever
possible. I failed to add a limit to libgsf due to error downloading
that tarball. There are so many of these that we'll just need to deal
with it in the future.

Also, fix udisks module which did not have any corresponding
gnome-build-meta element due to a naming difference.