Remove libinput and udisks from tarball conversion
libinput is no longer doing tarball releases, only GitLab releases, so
it doesn't belong here anymore.

I fixed udisks to work in the previous commit, whereupon I discovered
that the latest udisks tarball is very old, which we could not have
noticed due to the naming error. So remove this one too.