Remove links to wiki
wiki.gnome.org is being decommissioned.

Of course, most of these links are actually only used when generating
the schedule wiki page, which will no longer happen, but this is step
one to replacing that with something newer. And one of the links is used
in our release reminders, which is definitely worth replacing now.