tools: Rename version files after the GNOME version
Once upon a time we only had version and version-stable,
easy to manage and keep track.

However it gets complicated when we branch now that we
support more than a single version at a time, as we
have to replace the old version-stable file with the
new one, while keeping it around for the rest of the
update cycle.

Thus name the files after the versions as its simpler
to manage. It also matches the tarball-config files
better.

Also add back the version file for 48.4 that was
replaced in 1844f8d7